#44 fix : 기타 api response fix

+++ app/src/main/java/com/aoztg/greengrim/data/model/response/HomeInfoResponse.kt
@@ -3,5 +3,5 @@ package com.aoztg.greengrim.data.model.response
 data class HomeInfoResponse(
     val id: Int,
     val nickName: String,
-    val carbonReduction: Int
+    val carbonReduction: Float
 )
